Add wall tiles, separation and central config
Introduce a central simulation config (config.h/cpp) and migrate per-processor config/params to getters. Add wall tile support with tile encoding/decoding, in-engine storage, middle-mouse placement and rendering in Game; implement wall collision processor (walls.*) and wire it into sim tick. Extract separation logic into its own processor and adjust avoidance/steering/arrival APIs to use dt and central configs. Change arrival tracking from a bool/byte to a float countdown for bump delays. Update UI to expose new config panels and tweak layout, and update CMakeLists to include the new sources.

+++ src/simulation/agent.h
@@ -6,7 +6,7 @@ struct AgentData {
     std::vector<Vector3>  positions;
     std::vector<Vector3>  targets;
     std::vector<Vector3>  vel;
-    std::vector<uint8_t>  arrived;
+    std::vector<float>    arrived;
 
     int size() const { return (int)positions.size(); }
 
@@ -14,6 +14,6 @@ struct AgentData {
         positions.push_back(pos);
         targets.push_back(pos);
         vel.push_back({ 0.0f, 0.0f, 0.0f });
-        arrived.push_back(0);
+        arrived.push_back(0.0f);
     }
 };
